Add NJsonSchema.Annotations.csproj
