+++ src/NJsonSchema.Annotations/IJsonSchemaExtensionDataAttribute.cs
@@ -8,12 +8,11 @@
 
 using System.Collections.Generic;
 
-namespace NJsonSchema.Annotations
+namespace NJsonSchema.Annotations;
+
+/// <summary>Interface to add an extension data property to a class or property, implementation needs to inherit from System.Attribute.</summary>
+public interface IJsonSchemaExtensionDataAttribute
 {
-    /// <summary>Interface to add an extension data property to a class or property, implementation needs to inherit from System.Attribute.</summary>
-    public interface IJsonSchemaExtensionDataAttribute
-    {
-        /// <summary>Gets the extension data.</summary>
-        IReadOnlyDictionary<string, object> ExtensionData { get; }
-    }
+    /// <summary>Gets the extension data.</summary>
+    IReadOnlyDictionary<string, object> ExtensionData { get; }
 }
