+++ src/NJsonSchema.sln
@@ -46,6 +46,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "01 Core", "01 Core", "{CDBC
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "02 CodeGeneration", "02 CodeGeneration", "{62F296F1-8CE9-493C-9022-35F4C687460B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NJsonSchema.Annotations", "NJsonSchema.Annotations\NJsonSchema.Annotations.csproj", "{526020E1-D3B5-49F6-BE11-5F4402A02E92}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -274,6 +276,22 @@ Global
 		{AFA1E1E7-37F9-4958-B0E3-ADB12F53E563}.Release|x64.Build.0 = Release|Any CPU
 		{AFA1E1E7-37F9-4958-B0E3-ADB12F53E563}.Release|x86.ActiveCfg = Release|Any CPU
 		{AFA1E1E7-37F9-4958-B0E3-ADB12F53E563}.Release|x86.Build.0 = Release|Any CPU
+		{526020E1-D3B5-49F6-BE11-5F4402A02E92}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{526020E1-D3B5-49F6-BE11-5F4402A02E92}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{526020E1-D3B5-49F6-BE11-5F4402A02E92}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{526020E1-D3B5-49F6-BE11-5F4402A02E92}.Debug|ARM.Build.0 = Debug|Any CPU
+		{526020E1-D3B5-49F6-BE11-5F4402A02E92}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{526020E1-D3B5-49F6-BE11-5F4402A02E92}.Debug|x64.Build.0 = Debug|Any CPU
+		{526020E1-D3B5-49F6-BE11-5F4402A02E92}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{526020E1-D3B5-49F6-BE11-5F4402A02E92}.Debug|x86.Build.0 = Debug|Any CPU
+		{526020E1-D3B5-49F6-BE11-5F4402A02E92}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{526020E1-D3B5-49F6-BE11-5F4402A02E92}.Release|Any CPU.Build.0 = Release|Any CPU
+		{526020E1-D3B5-49F6-BE11-5F4402A02E92}.Release|ARM.ActiveCfg = Release|Any CPU
+		{526020E1-D3B5-49F6-BE11-5F4402A02E92}.Release|ARM.Build.0 = Release|Any CPU
+		{526020E1-D3B5-49F6-BE11-5F4402A02E92}.Release|x64.ActiveCfg = Release|Any CPU
+		{526020E1-D3B5-49F6-BE11-5F4402A02E92}.Release|x64.Build.0 = Release|Any CPU
+		{526020E1-D3B5-49F6-BE11-5F4402A02E92}.Release|x86.ActiveCfg = Release|Any CPU
+		{526020E1-D3B5-49F6-BE11-5F4402A02E92}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -293,6 +311,7 @@ Global
 		{3D8BA0A7-65AF-4328-82DC-8C131D9A9EA8} = {62F296F1-8CE9-493C-9022-35F4C687460B}
 		{E767A007-6007-4898-B80A-FE4ACBF2C588} = {62F296F1-8CE9-493C-9022-35F4C687460B}
 		{4F0A74F2-54BE-4747-8EDB-562ABB6E0B64} = {CDBC7AC4-9D3D-430D-86BF-CDAECBDD715A}
+		{526020E1-D3B5-49F6-BE11-5F4402A02E92} = {CDBC7AC4-9D3D-430D-86BF-CDAECBDD715A}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {9D5EDC80-5611-493B-804B-8B364816952B}
